refactor: processes, process-list, process-icon-view
