+++ src/management-system-v2/components/item-list-view.module.scss
@@ -13,3 +13,12 @@
 .TitleTruncation {
   max-width: 20dvw;
 }
+
+.HoverableTableCell:not(th) {
+  opacity: 0;
+  transition: opacity 0.2s ease !important;
+}
+
+tr:has(.HoverableTableCell):hover .HoverableTableCell {
+  opacity: 1;
+}
